Overworld, Scene linking
1. Controller support is currently being achieved
a. The main menu now works for controller
b. There is now a dropdown menu in the overworld
c. It works by pressing the North Button (triangle)
d. The overworld movement currently only works for controller
2. Removed Unnecessary Scenes.
a. Single Player Scenes are being worked on, So multiplayer and practice scenes have been deleted to cut down on file size.

TODO: Work on getting the battle system to be complient with the controller. Work on overworld movement options i.e jumping, air blasts etc.
Work on trying to make the scenes load faster.